Update index.md
Changed @when references to after_wp_load.  I think the propensity for package creators will be to assume WP functionality.  If they copy and paste command parameters this will provide that default starting point.

Changed text under @when heading to indicate it supports all WP-CLI hooks.  Again, I think this is defensive wording for the devs that logically approach this from the plugin world. 

Finally, added authors example array to the composer.json file near the end of the page.