Add link to WP-CLI Support. Edit text for clarity.
* Capitalize sentences in bullet points

* There was a note that `5.x is only supported when running php7, phpunit 4.8 is required when running php5` and `**Note**: phpunit 4.8.x is required (5.x won't work)` These seem confusing and a little contradictory so I removed the second note and kept the first which give better compatibility info
* Running the install script had 2 commands:
```
cd $(wp plugin path my-plugin --dir)
bash bin/install-wp-tests.sh wordpress_test root '' localhost latest
```
I specified that you need to be in the plugin directory and removed the `cd $(wp plugin path my-plugin --dir)` line. Although cool and interesting to use a sub shell command, it's a less known syntax and it was hard to understand what was going on. Seemed to break the rule of clarity over cleverness.

I also moved the explainer text around a little so it is easier to understand that we are explaining the parameters shown, not that you have to replace those (kinda generic looking params) with your own.

* For bash commands on their own line, I added the 3 ticks and specified bash so it will show proper color coding.

* Changed references for `tmp/` to `/tmp` to make it clearer that this is off the root directory.

* Added a note that if you have trouble, check out the WP-CLI Support section. This was the impetus for my edit. I had trouble running tests while using XAMPP. I was getting a "Can’t connect to the database" error and searching all around the internet for an answer. This (https://make.wordpress.org/cli/handbook/common-issues/) answer for using a custom PHP library solved it but wasn't easy to find. Since lots of people use MAMP or XAMPP, I considered how to include this information, but a link to this existing documentation is the best solution so the answer is maintained in a single place. Also, never hurts to point out where to get help.