Update to clarify instructions
I'm new to unit testing and if I'd known I didn't have to set up a server to generate the test files I wouldn't have bothered. So hopefully this clarifies that there's not much to it if you want to run unit tests on Travis. Let me know if I've missed anything.