Add lima-city to hosting companies
This change adds lima-city.de to the list of hosting companies that have cp-cli pre-installed. wp-cli is preinstalled on all hosting plans that offer SSH.

We also have a web interface for managing plugins, themes and updates. If you need a demo to verify the claim just say the word.

Have a nice day.